+++ src/lib.rs
@@ -63,7 +63,7 @@
 //!   into an allocation using [`Marker::allocate_slice()`],
 //!   [`Marker::allocate_slice_clone()`], or
 //!   [`Marker::allocate_slice_copy()`], and existing allocations can be
-//!   converted into slice allocations using the [`IntoSliceAllocation`]
+//!   converted into slice allocations using the [`IntoSliceLikeAllocation`]
 //!   trait. Allocations of [`str`] slices are supported as well.
 //! - **Extending and concatenating allocations.** Under certain conditions,
 //!   allocations can be added to or combined:
@@ -443,7 +443,7 @@
 //! [`CACHE_ALIGNMENT`]: constant.CACHE_ALIGNMENT.html
 //! [`CacheAligned`]: struct.CacheAligned.html
 //! [`Drop`]: https://doc.rust-lang.org/core/ops/trait.Drop.html
-//! [`IntoSliceAllocation`]: trait.IntoSliceAllocation.html
+//! [`IntoSliceLikeAllocation`]: trait.IntoSliceLikeAllocation.html
 //! [`mark_back()`]: struct.Scratchpad.html#method.mark_back
 //! [`mark_front()`]: struct.Scratchpad.html#method.mark_front
 //! [`Marker`]: trait.Marker.html