+++ src/traits.rs
@@ -439,61 +439,71 @@ pub trait ConcatenateSlice: SliceLike {}
 impl<T> ConcatenateSlice for [T] {}
 impl ConcatenateSlice for str {}
 
-/// Trait for safely converting an [`Allocation`] of a given type into a slice
-/// allocation without losing or altering any of the allocation data.
+/// Trait for safely converting an [`Allocation`] of a given type into a
+/// [`SliceLike`] allocation without losing or altering any of the allocation
+/// data.
 ///
-/// The [`into_slice_allocation()`] and [`into_element_slice_allocation()`]
-/// methods perform permanent conversions of an allocation into a slice
-/// allocation. For temporary access to the contents of an allocation as a
-/// slice, the [`as_slice()`], [`as_mut_slice()`], [`as_slice_ptr()`], and
-/// [`as_mut_slice_ptr()`] methods are also available.
+/// The [`into_slice_like_allocation()`] and
+/// [`into_element_slice_allocation()`] methods perform permanent conversions
+/// of an allocation into a slice allocation. For temporary access to the
+/// contents of an allocation as a slice, the [`as_slice_like()`],
+/// [`as_mut_slice_like()`], [`as_slice_like_ptr()`], and
+/// [`as_mut_slice_like_ptr()`] methods are also available.
 ///
 /// # Safety
 ///
 /// Implementing this trait implies that a type can be safely consumed when
-/// converting into a slice allocation. Implementing it for other types can
-/// cause undefined behavior or crashing.
+/// converting into a slice allocation, specifically the [`Allocation`] types
+/// for which it is already implemented by this crate. Incorrectly
+/// implementing it for other types can result in undefined behavior or
+/// runtime instability.
 ///
 /// [`Allocation`]: struct.Allocation.html
-/// [`as_mut_slice()`]: #method.as_mut_slice
-/// [`as_mut_slice_ptr()`]: #tymethod.as_mut_slice_ptr
-/// [`as_slice()`]: #method.as_slice
-/// [`as_slice_ptr()`]: #tymethod.as_slice_ptr
-/// [`into_slice_allocation()`]: #method.into_slice_allocation
-pub unsafe trait IntoSliceAllocation<'marker, T>: Sized
+/// [`as_mut_slice_like()`]: #tymethod.as_mut_slice_like
+/// [`as_mut_slice_like_ptr()`]: #method.as_mut_slice_like_ptr
+/// [`as_slice_like()`]: #tymethod.as_slice_like
+/// [`as_slice_like_ptr()`]: #method.as_slice_like_ptr
+/// [`into_slice_like_allocation()`]: #method.into_slice_like_allocation
+/// [`into_element_slice_allocation()`]: #method.into_element_slice_allocation
+/// [`SliceLike`]: trait.SliceLike.html
+pub unsafe trait IntoSliceLikeAllocation<'marker, T>: Sized
 where
     T: SliceLike + ?Sized,
 {
-    /// Converts this allocation into a slice allocation without altering the
-    /// allocation data.
+    /// Converts this allocation into an allocation of a compatible
+    /// [`SliceLike`] type without altering the allocation data.
     ///
     /// # Examples
     ///
     /// ```
-    /// use scratchpad::{Allocation, IntoSliceAllocation, Scratchpad};
+    /// use scratchpad::{Allocation, IntoSliceLikeAllocation, Scratchpad};
     ///
     /// let scratchpad = Scratchpad::<[i32; 6], [usize; 1]>::static_new();
     /// let marker = scratchpad.mark_front().unwrap();
     ///
     /// let scalar = marker.allocate(3).unwrap();
-    /// assert_eq!(*scalar.into_slice_allocation(), [3]);
+    /// assert_eq!(*scalar.into_slice_like_allocation(), [3]);
     ///
     /// let slice = marker.allocate_array_with(2, |index| index as i32)
     ///     .unwrap();
-    /// assert_eq!(*slice.into_slice_allocation(), [0, 1]);
+    /// assert_eq!(*slice.into_slice_like_allocation(), [0, 1]);
     ///
     /// // Automatic conversion of an array into a slice is ambiguous, as the
     /// // compiler can't tell whether we want a slice with the same length as
     /// // the array or a slice with only one element containing the entire
     /// // array. We must explicitly specify the slice type in this example.
     /// let array = marker.allocate([9, 8, 7]).unwrap();
-    /// let array_slice: Allocation<[i32]> = array.into_slice_allocation();
+    /// let array_slice: Allocation<[i32]> = array
+    ///     .into_slice_like_allocation();
     /// assert_eq!(*array_slice, [9, 8, 7]);
     /// ```
+    ///
+    /// [`SliceLike`]: trait.SliceLike.html
     #[inline]
-    fn into_slice_allocation(mut self) -> Allocation<'marker, T> {
+    fn into_slice_like_allocation(mut self) -> Allocation<'marker, T> {
         unsafe {
-            let slice_ptr = (*self.as_mut_slice_ptr()).as_element_slice_mut();
+            let slice_ptr =
+                (*self.as_mut_slice_like_ptr()).as_element_slice_mut();
             let ptr = (*slice_ptr).as_mut_ptr();
             let len = (*slice_ptr).len();
             forget(self);
@@ -507,13 +517,13 @@ where
         }
     }
 
-    /// Converts this allocation into a [`[T]`] slice allocation without
-    /// altering the allocation data.
+    /// Converts this allocation into an allocation of elements that make up
+    /// a compatible [`SliceLike`] type without altering the allocation data.
     ///
     /// # Examples
     ///
     /// ```
-    /// use scratchpad::{IntoSliceAllocation, Scratchpad};
+    /// use scratchpad::{IntoSliceLikeAllocation, Scratchpad};
     ///
     /// let scratchpad = Scratchpad::<[u8; 32], [usize; 1]>::static_new();
     /// let marker = scratchpad.mark_front().unwrap();
@@ -527,12 +537,13 @@ where
     /// assert_eq!(*bytes, [b'f', b'o', b'o']);
     /// ```
     ///
-    /// [`[T]`]: https://doc.rust-lang.org/std/primitive.slice.html
+    /// [`SliceLike`]: trait.SliceLike.html
     fn into_element_slice_allocation(
         mut self,
     ) -> Allocation<'marker, [<T as SliceLike>::Element]> {
         unsafe {
-            let slice_ptr = (*self.as_mut_slice_ptr()).as_element_slice_mut();
+            let slice_ptr =
+                (*self.as_mut_slice_like_ptr()).as_element_slice_mut();
             let ptr = (*slice_ptr).as_mut_ptr();
             let len = (*slice_ptr).len();
             forget(self);
@@ -545,24 +556,24 @@ where
         }
     }
 
-    /// Returns a reference to a slice of this allocation's data without
-    /// altering the allocation.
+    /// Returns a reference to this allocation's data as a [`SliceLike`] type
+    /// without altering the allocation.
     ///
     /// # Examples
     ///
     /// ```
-    /// use scratchpad::{Allocation, IntoSliceAllocation, Scratchpad};
+    /// use scratchpad::{Allocation, IntoSliceLikeAllocation, Scratchpad};
     ///
     /// let scratchpad = Scratchpad::<[i32; 6], [usize; 1]>::static_new();
     /// let marker = scratchpad.mark_front().unwrap();
     ///
     /// let scalar = marker.allocate(3).unwrap();
-    /// assert_eq!(*scalar.as_slice(), [3]);
+    /// assert_eq!(*scalar.as_slice_like(), [3]);
     /// assert_eq!(*scalar, 3);
     ///
     /// let slice = marker.allocate_array_with(2, |index| index as i32)
     ///     .unwrap();
-    /// assert_eq!(*slice.as_slice(), [0, 1]);
+    /// assert_eq!(*slice.as_slice_like(), [0, 1]);
     /// assert_eq!(*slice, [0, 1]);
     ///
     /// // Automatic conversion of an array into a slice is ambiguous, as the
@@ -570,65 +581,21 @@ where
     /// // the array or a slice with only one element containing the entire
     /// // array. We must explicitly specify the slice type in this example.
     /// let array = marker.allocate([9, 8, 7]).unwrap();
-    /// let array_slice: &[i32] = array.as_slice();
+    /// let array_slice: &[i32] = array.as_slice_like();
     /// assert_eq!(*array_slice, [9, 8, 7]);
     /// assert_eq!(*array, [9, 8, 7]);
     /// ```
-    #[inline]
-    fn as_slice(&self) -> &T {
-        unsafe { &*self.as_slice_ptr() }
-    }
+    ///
+    /// [`SliceLike`]: trait.SliceLike.html
+    fn as_slice_like(&self) -> &T;
 
-    /// Returns a mutable reference to a slice of this allocation's data
-    /// without altering the allocation.
+    /// Returns a mutable reference to this allocation's data as a
+    /// [`SliceLike`] type without altering the allocation.
     ///
     /// # Examples
     ///
     /// ```
-    /// use scratchpad::{Allocation, IntoSliceAllocation, Scratchpad};
-    ///
-    /// let scratchpad = Scratchpad::<[i32; 3], [usize; 1]>::static_new();
-    /// let marker = scratchpad.mark_front().unwrap();
-    ///
-    /// let mut array = marker.allocate([9, 8, 7]).unwrap();
-    ///
-    /// // Restrict scope of the mutable borrow performed by `as_mut_slice()`.
-    /// {
-    ///     let array_slice: &mut [i32] = array.as_mut_slice();
-    ///     array_slice[1] = 4;
-    /// }
-    ///
-    /// assert_eq!(*array, [9, 4, 7]);
-    /// ```
-    #[inline]
-    fn as_mut_slice(&mut self) -> &mut T {
-        unsafe { &mut *self.as_mut_slice_ptr() }
-    }
-
-    /// Returns a pointer to a slice of this allocation's data without
-    /// altering the allocation.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use scratchpad::{Allocation, IntoSliceAllocation, Scratchpad};
-    ///
-    /// let scratchpad = Scratchpad::<[i32; 3], [usize; 1]>::static_new();
-    /// let marker = scratchpad.mark_front().unwrap();
-    ///
-    /// let array = marker.allocate([9, 8, 7]).unwrap();
-    /// let array_slice: *const [i32] = array.as_slice_ptr();
-    /// assert_eq!(*unsafe { &*array_slice }, [9, 8, 7]);
-    /// ```
-    fn as_slice_ptr(&self) -> *const T;
-
-    /// Returns a mutable pointer to a slice of this allocation's data without
-    /// altering the allocation.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use scratchpad::{Allocation, IntoSliceAllocation, Scratchpad};
+    /// use scratchpad::{Allocation, IntoSliceLikeAllocation, Scratchpad};
     ///
     /// let scratchpad = Scratchpad::<[i32; 3], [usize; 1]>::static_new();
     /// let marker = scratchpad.mark_front().unwrap();
@@ -636,9 +603,57 @@ where
     /// let mut array = marker.allocate([9, 8, 7]).unwrap();
     ///
     /// // Restrict scope of the mutable borrow performed by
-    /// // `as_mut_slice_ptr()`.
+    /// // `as_mut_slice_like()`.
     /// {
-    ///     let array_slice: *mut [i32] = array.as_mut_slice_ptr();
+    ///     let array_slice: &mut [i32] = array.as_mut_slice_like();
+    ///     array_slice[1] = 4;
+    /// }
+    ///
+    /// assert_eq!(*array, [9, 4, 7]);
+    /// ```
+    ///
+    /// [`SliceLike`]: trait.SliceLike.html
+    fn as_mut_slice_like(&mut self) -> &mut T;
+
+    /// Returns a pointer to this allocation's data as a [`SliceLike`] type
+    /// without altering the allocation.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use scratchpad::{Allocation, IntoSliceLikeAllocation, Scratchpad};
+    ///
+    /// let scratchpad = Scratchpad::<[i32; 3], [usize; 1]>::static_new();
+    /// let marker = scratchpad.mark_front().unwrap();
+    ///
+    /// let array = marker.allocate([9, 8, 7]).unwrap();
+    /// let array_slice: *const [i32] = array.as_slice_like_ptr();
+    /// assert_eq!(*unsafe { &*array_slice }, [9, 8, 7]);
+    /// ```
+    ///
+    /// [`SliceLike`]: trait.SliceLike.html
+    #[inline]
+    fn as_slice_like_ptr(&self) -> *const T {
+        self.as_slice_like()
+    }
+
+    /// Returns a mutable pointer to this allocation's data as a [`SliceLike`]
+    /// type without altering the allocation.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use scratchpad::{Allocation, IntoSliceLikeAllocation, Scratchpad};
+    ///
+    /// let scratchpad = Scratchpad::<[i32; 3], [usize; 1]>::static_new();
+    /// let marker = scratchpad.mark_front().unwrap();
+    ///
+    /// let mut array = marker.allocate([9, 8, 7]).unwrap();
+    ///
+    /// // Restrict scope of the mutable borrow performed by
+    /// // `as_mut_slice_like_ptr()`.
+    /// {
+    ///     let array_slice: *mut [i32] = array.as_mut_slice_like_ptr();
     ///     unsafe {
     ///         (*array_slice)[1] = 4;
     ///     }
@@ -646,37 +661,42 @@ where
     ///
     /// assert_eq!(*array, [9, 4, 7]);
     /// ```
-    fn as_mut_slice_ptr(&mut self) -> *mut T;
+    ///
+    /// [`SliceLike`]: trait.SliceLike.html
+    #[inline]
+    fn as_mut_slice_like_ptr(&mut self) -> *mut T {
+        self.as_mut_slice_like()
+    }
 }
 
-unsafe impl<'marker, T> IntoSliceAllocation<'marker, T>
+unsafe impl<'marker, T> IntoSliceLikeAllocation<'marker, T>
     for Allocation<'marker, T>
 where
     T: SliceLike + ?Sized,
 {
     #[inline]
-    fn as_slice_ptr(&self) -> *const T {
-        self.data.as_ptr()
+    fn as_slice_like(&self) -> &T {
+        unsafe { self.data.as_ref() }
     }
 
     #[inline]
-    fn as_mut_slice_ptr(&mut self) -> *mut T {
-        self.data.as_ptr()
+    fn as_mut_slice_like(&mut self) -> &mut T {
+        unsafe { self.data.as_mut() }
     }
 }
 
-unsafe impl<'marker, T> IntoSliceAllocation<'marker, [T]>
+unsafe impl<'marker, T> IntoSliceLikeAllocation<'marker, [T]>
     for Allocation<'marker, T>
 where
     T: Sized,
 {
     #[inline]
-    fn as_slice_ptr(&self) -> *const [T] {
+    fn as_slice_like(&self) -> &[T] {
         unsafe { slice::from_raw_parts(self.data.as_ptr(), 1) }
     }
 
     #[inline]
-    fn as_mut_slice_ptr(&mut self) -> *mut [T] {
+    fn as_mut_slice_like(&mut self) -> &mut [T] {
         unsafe { slice::from_raw_parts_mut(self.data.as_ptr(), 1) }
     }
 }
@@ -1327,19 +1347,19 @@ macro_rules! generate_array_trait_impls {
             }
         }
 
-        unsafe impl<'marker, T> IntoSliceAllocation<'marker, [T]>
+        unsafe impl<'marker, T> IntoSliceLikeAllocation<'marker, [T]>
             for Allocation<'marker, [T; $size]>
         where
             T: Sized,
         {
             #[inline]
-            fn as_slice_ptr(&self) -> *const [T] {
-                unsafe { &(*self.data.as_ref())[..] }
+            fn as_slice_like(&self) -> &[T] {
+                unsafe { &self.data.as_ref()[..] }
             }
 
             #[inline]
-            fn as_mut_slice_ptr(&mut self) -> *mut [T] {
-                unsafe { &mut (*self.data.as_mut())[..] }
+            fn as_mut_slice_like(&mut self) -> &mut [T] {
+                unsafe { &mut self.data.as_mut()[..] }
             }
         }
     };